web-animations/animation-timing/animation-timing-duration.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationTiming duration attribute test</title>
<meta name="assert" content="The iteration duration which is a real number greater than or equal to zero (including positive infinity) representing the time taken to complete a single iteration of the animation node.">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationtiming-duration">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function createAnimationsToTest(test, timing) {
    var animation1 = new Animation(createDiv(test), KEYFRAMES, timing);
    var animation2 = new Animation(createDiv(test), KEYFRAMES,
        {iterations: Infinity, duration: ANIMATION_DURATION});
    var animation3 = new Animation(createDiv(test), KEYFRAMES,
        {iterations: Infinity, duration: ANIMATION_DURATION});
    var animationGroup = new AnimationGroup([animation2], timing);
    var animationSequence = new AnimationSequence([animation3], timing);
    return [animation1, animation2, animation3];
}

async_test(function(t) {
    var timing = {
        duration: ANIMATION_DURATION
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);

        assert_equals(animation.computedTiming.duration, ANIMATION_DURATION,
            'Wrong AnimationNode.computedTiming.iterations value');

        player.ready.then(t.step_func(function() {
            player.currentTime = ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target), ANIMATION_TOP_1,
                'Animation ' + type(node) + ' should be at 1');
            player.currentTime = ANIMATION_DURATION + 1;
            assert_equals(window.getComputedStyle(animation.target), ANIMATION_TOP_DEFAULT,
                'Animation ' + type(node) + ' should be finished');
            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that if value of AnimationTiming.duration attribute returns iteration duration ' +
    'of the animation node');
</script>
</body>
